refactor(config): move matcher/regex types to config/common
This moves MatchRegexps,Regexp and Matcher types to config/common. It
allows to reuse theses types in the notify package without creating
circular dependencies between the notify and config packages. This is a
preparation to further separate notifier specific code from the config
package and move it to the notify sub packages.
