+++ config/config_test.go
@@ -538,69 +538,6 @@ func TestJSONMarshal(t *testing.T) {
 	}
 }
 
-func TestMarshalRegexpWithNilValue(t *testing.T) {
-	r := &Regexp{}
-
-	out, err := json.Marshal(r)
-	require.NoError(t, err)
-	require.Equal(t, "null", string(out))
-
-	out, err = yaml.Marshal(r)
-	require.NoError(t, err)
-	require.Equal(t, "null\n", string(out))
-}
-
-func TestUnmarshalEmptyRegexp(t *testing.T) {
-	b := []byte(`""`)
-
-	{
-		var re Regexp
-		err := json.Unmarshal(b, &re)
-		require.NoError(t, err)
-		require.Equal(t, regexp.MustCompile("^(?:)$"), re.Regexp)
-		require.Empty(t, re.original)
-	}
-
-	{
-		var re Regexp
-		err := yaml.Unmarshal(b, &re)
-		require.NoError(t, err)
-		require.Equal(t, regexp.MustCompile("^(?:)$"), re.Regexp)
-		require.Empty(t, re.original)
-	}
-}
-
-func TestUnmarshalNullRegexp(t *testing.T) {
-	input := []byte(`null`)
-
-	{
-		var re Regexp
-		err := json.Unmarshal(input, &re)
-		require.NoError(t, err)
-		require.Empty(t, re.original)
-	}
-
-	{
-		var re Regexp
-		err := yaml.Unmarshal(input, &re) // Interestingly enough, unmarshalling `null` in YAML doesn't even call UnmarshalYAML.
-		require.NoError(t, err)
-		require.Nil(t, re.Regexp)
-		require.Empty(t, re.original)
-	}
-}
-
-func TestMarshalEmptyMatchers(t *testing.T) {
-	r := Matchers{}
-
-	out, err := json.Marshal(r)
-	require.NoError(t, err)
-	require.Equal(t, "[]", string(out))
-
-	out, err = yaml.Marshal(r)
-	require.NoError(t, err)
-	require.Equal(t, "[]\n", string(out))
-}
-
 func TestJSONUnmarshal(t *testing.T) {
 	c, err := LoadFile("testdata/conf.good.yml")
 	if err != nil {
@@ -656,9 +593,9 @@ receivers:
 
 func TestEmptyFieldsAndRegex(t *testing.T) {
 	boolFoo := true
-	regexpFoo := Regexp{
+	regexpFoo := amcommoncfg.Regexp{
 		Regexp:   regexp.MustCompile("^(?:^(foo1|foo2|baz)$)$"),
-		original: "^(foo1|foo2|baz)$",
+		Original: "^(foo1|foo2|baz)$",
 	}
 
 	expectedConf := Config{
@@ -704,7 +641,7 @@ func TestEmptyFieldsAndRegex(t *testing.T) {
 			Routes: []*Route{
 				{
 					Receiver: "team-X-mails",
-					MatchRE: map[string]Regexp{
+					MatchRE: map[string]amcommoncfg.Regexp{
 						"service": regexpFoo,
 					},
 				},
